Fairly major change to custom tables
The main issue I started off wanting to address was that the code
for a mean (sd) format of table was really complex, and needlessly
so given that this format is not in fact APA style.

Rounding was also handled in a needlessly complex way - kable allows
you to set number of decimal places directly.

Pedagogically, too much was attempted at each step. Never do more than
a couple of simple steps before stopping and explaining what you've
done, even if that results in the code being a few lines longer.