Revise introductory paragraphs
You were mainly on the right lines, but the text wasn't particularly
clear on the implications of the Central Limit Theorem, and needed
a clearer 'how to' on when it would be sensible to use these tests.

I also removed the bit about outliers. 'Outlier' is a vexed term
in psychology, sometimes used to mean 'comes from a different
distribution' (e.g. a participant who pressed one key throughout
and so wasn't doing the task), and sometimes used to mean
'is a relatively rare point from the distribution' (e.g. an IQ
of 130). If you meant the latter, then this is the same point
as already made about distribution shape, and needn't be made again.
If you meant the former, then such data points should ideally
be removed. We could probably have a whole worksheet on this,
including about the way 'outlier removal' under the second
scenario is part of what gave us the replication crisis.