Revise explanation of choice of median
The reasons we report medians in NP tests are different to the
reasons we report means in P tests.

In P tests, the mean feeds directly into the statistic (e.g. t)
so it would be odd to report anything else if that's the only
inferential test you used.

In NP tests, it's true that the median and the U are both based
on operations that involve ranking data, but it's not the case that
the median feeds into the calculation in quite the same way.

The reason we tend to report medians is as I've put in the text