Refer to Wilcoxon rank-sum rather than Mann-Whitney U
The command is called 'wilcoxon.test', so let's call it a Wilcoxon.
The name usually used is Wilcoxon rank-sum, to distinguish it from
the Wilcoxon signed-ranks - so let's call it that.

Added a section to explain that the two are equivalent.

I see you were a bit led astray by:
https://data.library.virginia.edu/the-wilcoxon-rank-sum-test/

and thought about including code to calculate the 'real' W
from wilcoxon.test. That page has a particularly historical
view on the whole thing (I guess not surprising for a library).
Their adjustment would give you the value as calculated by
Fred Wilcoxon in 1945, but certainly for as long as I've been doing
this (about 30 years, including by hand and using paper look-up
tables), W is defined in such a way that W = U.