Wilcoxon does not test for difference in medians
Although this is a common misconception, the Wilcoxon is _not_
a test for a difference in the medians. It's a test for what it
actually calculates! (see ?wilcoxon.test for details).

I guess we could get into this in the text - sometimes people say
that it is a test for whether the two distributions are different,
which is more accurate. But in this context I think it's easier
to just be vague.