Correct explanation of Kruskal-Wallis
The main thing I needed to correct here was the assertion that this
test was related to a chi-square test. It isn't really, such tests
use count data, and these are ordinal data. The link is that it's
possible to estimate the H statistic using a chi-square
_distribution_ with degrees of freedom set to rows-1. The H distribution
and the chi-square distribution are not the same, they're just similar
enough that it doesn't affect the answer much as long as all groups are
n >= 5.

I also made a few cosmetic changes.