Remove object model
I'd done this a bit already but now they're all gone.

The point here is that it seemed like the only reason you were putting
results of analyses into objects was so your Rmarkdown could call
those objects. Pedadogigcally that's not a good reason, and the code
is shorter and easier for students to follow if you don't use an object.

Generally speaking, don't use objects unless the student is later
going to use that object.