Major change to preprocessing
The preprocessing pipeline was pretty complex, using a bunch of
commands new to the reader, and involving defining functions.
It seemed to me this could be largely avoided by creating a single
long data frame as the first preprocessing setep.

Note: Also, the exclusion for poor performance seemed to be wrong
- you seemed to be excluding the best perfomers, rather than the worst.