Major changes to rest of worksheet
Here were the issues I attempted to resolve in this re-write:

(1) Poor practice (exclusions)

The only thing you achieve by dropping people with accuracy scores
below a certain cutoff based on mean/sd performance in this task is
to reduce variability and hence inflate the chances of a Type I
error. Don't use atheoretical performance-based subject exclusions
in this way.

(2) Overly complex code

There was quite a lot of writing a function and then applying
that function repeatedly. Also complex code to get inferential
stats into a descriptive table. Although I can see the appeal
of the former for advanced users, and the appeal of the latter
if you want to produce reproducible reports with Rmarkdown,
neither of those things apply to this relatively introductory
course. A certain amount of repetition, and of manually
transferring results of inferential tests from R output to
a wordprocessor document is preferrable for this audience,
in order to keep the code simpler.

(3) Poor practice 2

It makes no sense to report a Spearman's correlation (non-parametric)
and then go on to do an essentially parametric inferential test
(correlationBF). Report the correlation coeff that corresponds
to the test you are performing, which in this case is the
standard Pearson's.