Correct fairly major error
Education level had not been handled correctly. Factors are still
unordered variables, even if you set an ordering for their display.
As a result, `lm` turned each level of this factor into a separate
binary regressor. This is presumably not desired or intended.

The way to handle any ordinal variable is to turn it into a number.
I've corrected the code and associated explanations.