Simplify line graph
The stat_summary thing was kinda cool, so I kept that despite it
technically being an unneeded new pattern.

However, one should not be 'dodging' this line graph - there's no
need now it has no error bars, and it looks a bit odd.

I also replaced 'dodge' with 0.1 later in the script, partly
to get it to compile, but also because this should be a pick+mix
worksheet, i.e. code from one graph type should not depend on
code from a different graph type.