Changed my mind
I first thought that I would change the one-factor Perruchet
graph to us stat_summary, then I realised that this was not
straight forward and that this would leave us with two different
ways in the same worksheet to plot means.

So, I reverted to the calculating means separately approach.
This means we can also directly reference the worksheet where
they already learned this graph.