Tweak, explain two-factor w/subj
This was pretty good, but I felt some more general explanatory
text was needed. Plus, I thought it'd be a good idea to add the
mean score to the density plot.

Incidentally, this is what a lot of congruency-effect data is
really like -- effects in the order of 10s of ms, generally coming
from something like a 60/40 split of people showing the mean
result and its opposite.