Fairly major change to mixed factors plot
We talked about this before, but the plot with eight confidence
intervals is a really poor, misleading plot. On the plus side,
this just shows that it's not at all obvious why it's a bad plot,
which I guess is why one sees this sort of thing a lot in the
literature. So, I've used it in a section on why it's a bad plot.

This led to a bunch of knock-on changes, as I refocussed the section
on the sort of plot that is clearest for the hypothesis under
test.