Extend Exercise 2
Similar reasons as for Exercise 1, see git log.

However, also woth mentioning:

1. I took the view in both exercises that we didn't need to give
more practice in preprocessing here - so I just gave them the
code for that.

2. Exercise 2 averaged over trials. Actually, we do not need to
do that, and in some sense it is better that we do not, as it allows
the analysis function to make use of this information if it can.